feat: ExtractorNew => Extractor
